+++ path_planning.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <math.h> // abs()를 위해 포함
+#include <stdint.h> // uint8_t 정의
 
 #define ARRAY_CAPACITY 100 // 경로 및 빈자리 배열의 최대 크기 정의
 #define INF 99999
@@ -42,7 +43,7 @@ static int reconstruct_path(Node** nodes, int current_y, int current_x, int path
 }
 
 // ---------- A* 알고리즘 ----------
-int astar(const int* start_pos, const int* target_pos, int** map_matrix, int rows, int cols, int path[][2]) {
+int astar(const int* start_pos, const int* target_pos, uint8_t** map_matrix, int rows, int cols, int path[][2]) {
     int sy = start_pos[0];
     int sx = start_pos[1];
     int ty = target_pos[0];
@@ -121,7 +122,7 @@ int astar(const int* start_pos, const int* target_pos, int** map_matrix, int row
 // ---------- 주차 자리 선택 ----------
 void find_preferred_parking(
     const int* user_preference,
-    int** map_matrix,
+    uint8_t** map_matrix,
     int map_rows,
     int map_cols,
     int* goal
